Fix signed-int cycle-counter overflow that froze giant traces
The interconnect edge counters (DirectInterconnect/Bus/Mesh) were signed
int. They gate message movement on `edges % 2 == 1`; at 2^31 edges
(~1.07B cycles) the counter wrapped negative, making the modulo never 1,
so the interconnect stopped moving messages and the sim deadlocked
periodically (~1.07B and ~3.22B cycles). Widen the cycle/edge counters to
uint64_t so the modulo stays 0/1 and wraps cleanly.

Also widen CacheDataHandler m_cycle/m_ready_cycle and CPU
m_last_received_msg_cycle to uint64_t: the SPLASH giants exceed 2^32
cycles, where the old uint32_t stamps wrapped.

With this, raytrace/radiosity/radix run to completion across all
protocols (snoop + directory x MSI/MESI/MOESI).

+++ header/Interconnect/Bus.h
@@ -36,8 +36,9 @@ namespace octopus
         BusController* interconnect_controller;
         DebugPrint* dprint;
 
-        int m_bus_cycle;
-        int m_bus_cycle_edges;
+        uint64_t m_bus_cycle;
+        uint64_t m_bus_cycle_edges; // MUST be unsigned (see DirectInterconnect.h): signed
+                                    // overflow of the edge counter froze the bus on giant traces.
 
     public:
         Bus(ParametersMap map,
